updated dispatch server count method and moved alert string method to alerts service.
