+++ packages/core/src/services/event/service.ts
@@ -1,7 +1,7 @@
 import { AlertColor } from "../alert/types";
 import { Service as BaseService } from "../base-service/base-service";
 import { EErrorConditions, EErrorStatus } from "../health/types";
-import { INode, NodesModel } from "../../models";
+import { NodesModel } from "../../models";
 import {
   EAlertTypes,
   IRedisEvent,
@@ -9,7 +9,6 @@ import {
   IToggleServerParams,
 } from "./types";
 import { AlertTypes } from "../../types";
-import { s, is } from "../../utils";
 import env from "../../environment";
 
 export class Service extends BaseService {
@@ -124,24 +123,25 @@ export class Service extends BaseService {
     const node = await this.getNode(id);
     await NodesModel.updateOne({ _id: node.id }, { status, conditions });
     const { chain, backend, frontend, loadBalancers, dispatch, url } = node;
+    const pnfDispatch = env("PNF") && dispatch;
 
     const healthy = conditions === EErrorConditions.HEALTHY;
-    const notSynced =
-      env("PNF") && dispatch
-        ? conditions === EErrorConditions.NOT_SYNCHRONIZED ||
-          conditions === EErrorConditions.OFFLINE ||
-          conditions === EErrorConditions.NO_RESPONSE
-        : conditions === EErrorConditions.NOT_SYNCHRONIZED;
-    const dispatchFrontendDown = Boolean(env("PNF") && dispatch && frontend && notSynced);
+    const notSynced = pnfDispatch
+      ? conditions === EErrorConditions.NOT_SYNCHRONIZED ||
+        conditions === EErrorConditions.OFFLINE ||
+        conditions === EErrorConditions.NO_RESPONSE
+      : conditions === EErrorConditions.NOT_SYNCHRONIZED;
+    const dispatchFrontendDown = Boolean(pnfDispatch && frontend && notSynced);
 
     const nodeCount = await this.getServerCount({
       destination: frontend || backend,
       loadBalancers,
       frontendUrl: frontend ? url : null,
+      dispatch: pnfDispatch,
     });
-    const nodeTotal = await NodesModel.count({ chain: chain.id });
+    const nodeTotal = await NodesModel.count({ chain: chain.id, haProxy: true });
 
-    const { message, statusStr } = this.getAlertMessage(
+    const { message, statusStr } = this.alert.getAlertMessage(
       event,
       alertType,
       nodeCount,
@@ -191,7 +191,7 @@ export class Service extends BaseService {
         });
         const nodeTotal = await NodesModel.count({ chain: chain.id });
 
-        const { title, message } = this.getRotationMessage(
+        const { title, message } = this.alert.getRotationMessage(
           node,
           enable,
           "success",
@@ -204,7 +204,7 @@ export class Service extends BaseService {
         );
       }
     } catch (error) {
-      const { title, message } = this.getRotationMessage(
+      const { title, message } = this.alert.getRotationMessage(
         node,
         enable,
         "error",
@@ -224,85 +224,5 @@ export class Service extends BaseService {
       title: "URGENT ALERT! Dispatch Frontend is down!",
       details: ["Dispatchers' HAProxy frontend is down!", message].join("\n"),
     });
-  }
-
-  /* ----- Message String Methods ----- */
-  private getAlertMessage(
-    { count, conditions, name, ethSyncing, height, details }: IRedisEvent,
-    alertType: EAlertTypes,
-    nodeCount: number,
-    nodeTotal: number,
-    destination: string,
-  ): { message: string; statusStr: string } {
-    const badOracles = details?.badOracles?.join("\n");
-    const noOracle = details?.noOracle;
-
-    const statusStr = `${name} is ${conditions}.`;
-    const countStr =
-      alertType !== EAlertTypes.RESOLVED
-        ? `This event has occurred ${count} time${s(count)} since first occurrence.`
-        : "";
-    const ethSyncStr = ethSyncing ? `ETH Syncing - ${ethSyncing}` : "";
-    const heightStr = height
-      ? `Height - ${
-          typeof height === "number"
-            ? height
-            : `Internal: ${height.internalHeight} / External: ${height.externalHeight} / Delta: ${height.delta}`
-        }`
-      : "";
-    let nodeCountStr =
-      nodeCount && nodeCount >= 0
-        ? `${nodeCount} of ${nodeTotal} node${s(nodeTotal)} ${is(
-            nodeCount,
-          )} in rotation for ${destination}.`
-        : "";
-    if (nodeCount <= 1) nodeCountStr = `${nodeCountStr.toUpperCase()}`;
-    const badOracleStr = badOracles?.length
-      ? `\nWarning - Bad Oracle${s(badOracles.length)}\n${badOracles}`
-      : "";
-    const noOracleStr = noOracle
-      ? `\nWarning - No Oracle for node. Node has ${details?.numPeers} peers.`
-      : "";
-
-    return {
-      message: [countStr, ethSyncStr, heightStr, badOracleStr, noOracleStr]
-        .filter(Boolean)
-        .join("\n"),
-      statusStr,
-    };
-  }
-
-  private getRotationMessage(
-    { backend, chain, host, loadBalancers }: INode,
-    enable: boolean,
-    mode: "success" | "error",
-    nodeCount: number,
-    nodeTotal: number,
-    error?: any,
-  ): { title: string; message: string } {
-    const name = `${host.name}/${chain.name}`;
-    const haProxyMessage = this.getHAProxyMessage({
-      destination: backend,
-      loadBalancers,
-    });
-    const title = enable
-      ? {
-          success: `[Added] - Successfully added ${name} to rotation`,
-          error: `[Error] - Could not add ${name} to rotation`,
-        }[mode]
-      : {
-          success: `[Removed] - Successfully removed ${name} from rotation`,
-          error: `[Error] - Could not remove ${name} from rotation`,
-        }[mode];
-    let nodeCountStr =
-      nodeCount && nodeCount >= 0
-        ? `${nodeCount} of ${nodeTotal} node${s(nodeTotal)} ${is(
-            nodeCount,
-          )} in rotation for ${backend}.`
-        : "";
-    if (nodeCount <= 1) nodeCountStr = `${nodeCountStr.toUpperCase()}`;
-    const message = [haProxyMessage, nodeCountStr, error].filter(Boolean).join("\n");
-
-    return { title, message };
   }
 }
