+++ packages/core/src/services/haproxy/types.ts
@@ -2,4 +2,5 @@ export interface IHAProxyParams {
   destination: string;
   domain: string;
   server?: string;
+  dispatch?: boolean;
 }
